// src/enums/nameRequestStates.ts
export enum NameRequestStates {
  // API STATES
  APPROVED = 'APPROVED',
  CANCELLED = 'CANCELLED',
  COMPLETED = 'COMPLETED',
  CONDITIONAL = 'CONDITIONAL', // NR state
  CONDITION = 'CONDITION', // NR name state
  DRAFT = 'DRAFT',
  EXPIRED = 'EXPIRED',
  HISTORICAL = 'HISTORICAL',
  HOLD = 'HOLD',
  INPROGRESS = 'INPROGRESS',
  REJECTED = 'REJECTED',
  NRO_UPDATING = 'NRO_UPDATING',
  NO_CONTACT = 'NO_CONTACT',
  INCORRECT_CONTACT = 'INCORRECT_CONTACT',

  // ADDITIONAL UI STATES FOR ERROR HANDLING
  CONSUMED = 'CONSUMED',
  INCORRECT_CONTACT = 'INCORRECT_CONTACT',
  INVALID = 'INVALID',
  NEED_CONSENT = 'NEED_CONSENT',
  NO_CONTACT = 'NO_CONTACT',
  NOT_APPROVED = 'NOT_APPROVED',
  NOT_FOUND = 'NOT_FOUND',
}